+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Fixed respect start offset from topicoptions.WithReaderGetPartitionStartOffset for commit messages
 * Added `query.AllowImplicitSessions()` option for execute queries through `query.Client.{Exec,Query,QueryResultSet,QueryRow}` without explicit sessions
 
 ## v3.112.0