+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-* Added experimental config.WithDisableOptimisticUnban to disable fast node unban after pessimization
-* Fixed respect start offset from topicoptions.WithReaderGetPartitionStartOffset for commit messages
+* Added experimental `config.WithDisableOptimisticUnban` option to disable fast node unban after pessimization
+* Fixed respect start offset from `topicoptions.WithReaderGetPartitionStartOffset` for commit messages
 * Added `query.AllowImplicitSessions()` option for execute queries through `query.Client.{Exec,Query,QueryResultSet,QueryRow}` without explicit sessions
 
 ## v3.112.0